* origin/master: (112 commits)
  Use generic method correctly
  avoid creating proof obligations in ProofManagementDialog It is unnecessary to create proof obligations in most cases there and creating some proof obligations changes state shared with a large part of KeY (see #1715).
  JavaDoc for changed methods + Doubles for timeout
  Checkstyle/SonarQube
  Avoid reopening popup menu when selection button is pressed
  Bug fixes from the IdentityHashMap case study
  Solve missing plugin in shadowJar by adding `mergeServiceFiles()`
  SonarQube
  Checkstyle
  mark JSpinner backgrounds correctly when an error occurs
  Deactivate SMT button when no proof is loaded
  Check formatting  with Spotless
  Add JavaDoc
  Modify problems using solver sockets
  Minor changes
  checkstyle
  Write comments
  Delete solvers.txt (automatically created during build)
  Load solvers.txt from everywhere on classpath
  Fix failing legacy solver test cases by setting different SMT logic
  ...

+++ key/keyext.ui.testgen/src/main/java/de/uka/ilkd/key/gui/testgen/TestgenOptionsPanel.java
@@ -76,7 +76,7 @@ public class TestgenOptionsPanel extends SettingsPanel implements SettingsProvid
     private JSpinner getMaxProcesses() {
         return addNumberField("Concurrent processes:", 0, Integer.MAX_VALUE, 1, INFO_MAX_PROCESSES,
                 obj -> {
-                    settings.setConcurrentProcesses(obj);
+                    settings.setConcurrentProcesses(obj.intValue());
                     settings.fireSettingsChanged();
                 });
     }
@@ -84,7 +84,7 @@ public class TestgenOptionsPanel extends SettingsPanel implements SettingsProvid
     private JSpinner getMaxUnwinds() {
         return addNumberField("Maximal unwinds:", 0, Integer.MAX_VALUE, 1, INFO_MAX_UNWINDS,
                 e -> {
-                    settings.setMaxUnwinds(e);
+                    settings.setMaxUnwinds(e.intValue());
                     settings.fireSettingsChanged();
                 });
     }
